requirements.txt
south
psycopg2
django-annoying
django-extensions
